app/requirements.txt
Flask==2.3.2
flask-sqlalchemy==3.0.5
PyJWT==2.8.0
pytest==7.4.0
Werkzeug==2.3.7
json-log-formatter==0.5.2
prometheus-client==0.16.0
psutil==5.9.7
pytest-mock==3.10.0
flake8==6.0.0
black==23.1.0
